Assignment 2
This program reads data from 2 separate text files and calculates the sum, average, and letter grade. It will then write and output the results into a new file.
The program will then use a sorting algorithm to sort the data by grade average and print it to the screen. The program will also use a search algorithm to search data by student name.